:sparkles: feat: main page updated
