+++ components/play/item/DetailPlayItem.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import React from 'react'
-import NaturalImage from './NaturalImage'
-import * as Styled from './style'
+import NaturalImage from '../NaturalImage'
+import * as Styled from '../style'
 
 interface DetailPlayItem {
   mainImage: string
