+++ components/play/item/MainPlayItem.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import React from 'react'
-import * as Styled from './style'
+import * as Styled from '../style'
 
 const MainPlayItem = () => {
   return (
@@ -10,9 +10,12 @@ const MainPlayItem = () => {
         <br />
         아직도 직접 검색해?
         <br />
-        이제 케이크크에서 한번에!
+        이제 케이크크에서
+        <br /> 한번에!
       </h1>
-      <Link href="/">케이크 가게 찾으러 가기!</Link>
+      <Link href="/" className="description">
+        케이크 가게 찾으러 가기!
+      </Link>
     </Styled.MainPlayItem>
   )
 }
